requirements.txt
yfinance
pandas
numpy
mplfinance
requests
pandas_market_calendars
python-dotenv
